Changes CustomUserDetails -> UserAccountDetails classes
