+++ build.gradle
@@ -53,7 +53,6 @@ ext {
 
 dependencies {
     //compile time
-    //compile("org.hibernate:hibernate-validator:5.2.4.Final")
     compile("edu.pitt.dbmi:ccd-db:${ccdDBVersion}")
     compile("org.springframework.boot:spring-boot-starter-security:${springBootVersion}")
     compile("org.springframework.boot:spring-boot-starter-jdbc:${springBootVersion}")
